Fix config command
